Fix failed file deletion retries
fixes MAT-894
flag=none

[ignore-stage-results=Flakey Spec Catcher]

Test plan
- Using both local storage and s3 storage
  (and InstFS would be nice, but I don't
  know how to set that up locally)
- In destroy_content_and_replace (app/models/attachment.rb)
  set up a debugger before the Attachment::Storage line
- Delete a file using the API like this
  DELETE http://<canvas>/api/v1/files/:id?replace=true
- Raise an error that causes the file deletion to fail
  partway through
- Then delete the file again, but make sure it succeeds
  and verify that the file looks like the "file_removed"
  file and that the purgatory object looks like the
  previously removed file

Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/294835
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Gonzalo Penaranda <gonzalo.penaranda@instructure.com>
Product-Review: Mysti Lilla <mysti@instructure.com>

+++ app/models/attachment.rb
@@ -1582,19 +1582,18 @@ class Attachment < ActiveRecord::Base
   # object. It will replace the attachment content with a file_removed file.
   def destroy_content_and_replace(deleted_by_user = nil)
     shard.activate do
+      file_removed_path = self.class.file_removed_path
+      new_name = File.basename(file_removed_path)
       att = root_attachment_id? ? root_attachment : self
-      return true if Purgatory.where(attachment_id: att).active.exists?
+      return true if att.display_name == new_name
 
-      att.send_to_purgatory(deleted_by_user)
+      att.send_to_purgatory(deleted_by_user) unless Purgatory.where(attachment_id: att).active.exists?
       att.destroy_content
       att.thumbnail&.destroy
 
-      file_removed_path = self.class.file_removed_path
-      new_name = File.basename(file_removed_path)
-
       if att.instfs_hosted? && InstFS.enabled?
-        # dupliciate the base file_removed file to a unique uuid
-        att.instfs_uuid = InstFS.duplicate_file(self.class.file_removed_base_instfs_uuid)
+        # duplicate the base file_removed file to a unique uuid
+        att.instfs_uuid = InstFS.duplicate_file(Attachment.file_removed_base_instfs_uuid)
       else
         Attachments::Storage.store_for_attachment(att, File.open(file_removed_path))
       end
@@ -1637,7 +1636,7 @@ class Attachment < ActiveRecord::Base
   def send_to_purgatory(deleted_by_user = nil)
     make_rootless
     new_instfs_uuid = nil
-    if instfs_hosted? && InstFS.enabled? && instfs_uuid
+    if instfs_hosted? && InstFS.enabled?
       # copy to a new instfs file
       new_instfs_uuid = InstFS.duplicate_file(instfs_uuid)
     elsif Attachment.s3_storage? && s3object.exists?
@@ -1718,7 +1717,7 @@ class Attachment < ActiveRecord::Base
       self.instfs_uuid = nil
     elsif Attachment.s3_storage?
       s3object.delete unless ApplicationController.test_cluster?
-    else
+    elsif File.exist?(full_filename)
       FileUtils.rm full_filename
     end
   end
